requirements.txt
requests
python-dateutil
certifi